Fixed warning on MinGW in XYZ converter

+++ src/formats/rpgmaker/xyz_converter.cc
@@ -39,7 +39,7 @@ void XyzConverter::decode_internal(File &file) const
     const char *palette_indices = data.data() + 256 * 3;
     char *out = pixels.get();
 
-    for (size_t i = 0; i < width * height; i ++)
+    for (size_t i = 0; i < static_cast<size_t>(width * height); i ++)
     {
         size_t index = *palette_indices ++;
         *out ++ = palette[index * 3 + 0];
